pyproject.toml
[tool.poetry]
name = "tiralabra-connect4"
version = "0.1.0"
description = ""
authors = ["Rasmus Heikkinen <rasmus.heikkinen@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.23.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.4.2"
pylint = "^2.14.5"
poetry-pyinvoke-plugin = "^0.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
